yanger: Move ietf-ospf implementation to separate module
